+++ processor/src/test/resources/bad_input/com/example/component_state_ref/ParametersModel.java
@@ -1,0 +1,11 @@
+package com.example.component_state_ref;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.ComponentStateRef;
+
+@ArezComponent( allowEmpty = true )
+abstract class ParametersModel
+{
+  @ComponentStateRef
+  abstract boolean isReady( int i );
+}