+++ processor/src/test/resources/bad_input/com/example/component_state_ref/PrivateModel.java
@@ -1,0 +1,14 @@
+package com.example.component_state_ref;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.ComponentStateRef;
+
+@ArezComponent( allowEmpty = true )
+abstract class PrivateModel
+{
+  @ComponentStateRef
+  private boolean isReady()
+  {
+    return true;
+  }
+}