+++ processor/src/test/resources/input/com/example/component_state_ref/DisposingStateRefModel.java
@@ -1,0 +1,12 @@
+package com.example.component_state_ref;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.ComponentStateRef;
+import arez.annotations.State;
+
+@ArezComponent( allowEmpty = true )
+public abstract class DisposingStateRefModel
+{
+  @ComponentStateRef( State.DISPOSING )
+  protected abstract boolean isDisposing();
+}