+++ processor/src/test/resources/input/com/example/component_state_ref/ReadyStateRefModel.java
@@ -1,0 +1,12 @@
+package com.example.component_state_ref;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.ComponentStateRef;
+import arez.annotations.State;
+
+@ArezComponent( allowEmpty = true )
+public abstract class ReadyStateRefModel
+{
+  @ComponentStateRef( State.READY )
+  protected abstract boolean isReady();
+}